test(transforms/mdx): run the stub-export execution case from a file so Bun can load it
bun test rejects a long data: dynamic-import specifier with NameTooLong,
which failed the tests (bun) job. The case now writes the generated stub to
a temp file and imports it, and because that write is an on-disk effect it
moves from the colocated unit file to the stub-module integration suite.
Also re-lock the sanitizer opt-out baseline at 384 after the rebase.
