+++ tests/integration/transforms/mdx/stub-module-cache.test.ts
@@ -9,11 +9,44 @@ import {
   assertNotEquals,
   assertRejects,
   assertStringIncludes,
+  assertThrows,
 } from "#veryfront/testing/assert.ts";
 import { describe, it } from "#veryfront/testing/bdd.ts";
 import { makeTempDir, readTextFile, remove } from "#veryfront/testing/deno-compat.ts";
-import { toFileUrl } from "#veryfront/compat/path";
-import { createStubModule } from "#veryfront/transforms/mdx/esm-module-loader/utils/stub-module.ts";
+import { writeTextFile } from "#veryfront/compat/fs.ts";
+import { join, toFileUrl } from "#veryfront/compat/path";
+import {
+  createStubModule,
+  generateStubCode,
+} from "#veryfront/transforms/mdx/esm-module-loader/utils/stub-module.ts";
+
+describe("generateStubCode", () => {
+  it("generates named exports that throw a MissingModuleError when called", async () => {
+    // Evaluated from a real module file rather than a `data:` URL: Bun's test
+    // runner rejects a long `data:` dynamic-import specifier (NameTooLong),
+    // and the on-disk write is why this case lives in integration.
+    const tempDir = await makeTempDir({ prefix: "vf-stub-module-exec-" });
+
+    try {
+      const stubPath = join(tempDir, "stub.js");
+      await writeTextFile(stubPath, generateStubCode("/path/to/module.js", ["foo"]));
+      const stub = await import(toFileUrl(stubPath).href) as { foo: () => unknown };
+
+      const error = assertThrows(
+        () => stub.foo(),
+        Error,
+        "/path/to/module.js",
+      ) as Error;
+      assertEquals(
+        error.name,
+        "MissingModuleError",
+        "a named stub export must throw, not resolve to undefined",
+      );
+    } finally {
+      await remove(tempDir, { recursive: true });
+    }
+  });
+});
 
 describe("createStubModule", () => {
   it("writes a fail-on-import stub that throws the deferred error", async () => {
